requirement.txt
streamlit>=1.24.0
PyPDF2>=3.0.0
langchain>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.21.0
tqdm
rapidfuzz
python-dateutil
pandas
python-dotenv
